requirements.txt
openai==1.68.2
python-dotenv==1.1.0
Requests==2.32.3
